fix: resolve conversation
