+++ frontend/src/components/AvatarProfile.tsx
@@ -1,12 +1,11 @@
 import { useState, useEffect, useCallback } from "react";
 import { NetworkIDEnum, Social } from "@builddao/near-social-js";
 import { ViewMethod } from "../hooks/near-method";
+import { replaceIpfsUrl } from "../utils/ipfs";
 
 // Constants
 const DEFAULT_AVATAR =
   "https://ipfs.near.social/ipfs/bafkreibiyqabm3kl24gcb2oegb7pmwdi6wwrpui62iwb44l7uomnn3lhbi";
-const IPFS_REGEX =
-  /^(?:https?:\/\/)(?:[^\/]+\/ipfs\/)?(Qm[1-9A-HJ-NP-Za-km-z]{44,}|b[A-Za-z2-7]{58,}|B[A-Z2-7]{58,}|z[1-9A-HJ-NP-Za-km-z]{48,}|F[0-9A-F]{50,})(?:\.[^\/]+)?(\/.*)?$/g;
 
 interface NFTMetadata {
   base_uri?: string;
@@ -51,15 +50,11 @@ export const AvatarProfile: React.FC<AvatarProfileProps> = ({
   const replaceIpfs = useCallback(
     (imageUrl: string) => {
       if (oldUrl !== imageUrl && imageUrl) {
-        IPFS_REGEX.lastIndex = 0;
-        const match = IPFS_REGEX.exec(imageUrl);
-        if (match) {
-          const newImageUrl = `https://ipfs.near.social/ipfs/${match[1]}${match[2] || ""}`;
-          if (newImageUrl !== imageUrl) {
-            setOldUrl(imageUrl);
-            setImageUrl(newImageUrl);
-            return;
-          }
+        const newImageUrl = replaceIpfsUrl(imageUrl);
+        if (newImageUrl) {
+          setOldUrl(imageUrl);
+          setImageUrl(newImageUrl);
+          return;
         }
       }
       if (imageUrl !== null) {
