+++ frontend/src/components/providers/NearWalletProvider.tsx
@@ -3,6 +3,8 @@ import { WalletSelectorProvider } from "@near-wallet-selector/react-hook";
 import { setupMyNearWallet } from "@near-wallet-selector/my-near-wallet";
 import { setupMeteorWallet } from "@near-wallet-selector/meteor-wallet";
 import { setupHereWallet } from "@near-wallet-selector/here-wallet";
+import { setupBitteWallet } from "@near-wallet-selector/bitte-wallet";
+import { setupIntearWallet } from "intear-wallet";
 
 export function NearWalletProvider({
   children,
@@ -16,6 +18,8 @@ export function NearWalletProvider({
           setupMyNearWallet() as any,
           setupHereWallet(),
           setupMeteorWallet(),
+          setupBitteWallet() as any,
+          setupIntearWallet(),
         ],
         network: (process.env.PUBLIC_NETWORK || "testnet") as NetworkId,
         createAccessKeyFor: process.env.PUBLIC_SMART_CONTRACT,
